+++ packages/ui-deprecated/components/ui/tx/action-view.tsx
@@ -18,6 +18,7 @@ import { PositionOpenComponent } from './actions-views/position-open.tsx';
 import { PositionCloseComponent } from './actions-views/position-close.tsx';
 import { PositionWithdrawComponent } from './actions-views/position-withdraw.tsx';
 import { IbcRelayComponent } from './actions-views/ibc-relay.tsx';
+import { LiquidityTournamentVoteComponent } from './actions-views/liquidity-tournament-vote.tsx';
 
 type Case = Exclude<ActionView['actionView']['case'], undefined>;
 
@@ -47,6 +48,7 @@ const CASE_TO_LABEL: Record<Case, string> = {
   communityPoolDeposit: 'Community Pool Deposit',
   communityPoolOutput: 'Community Pool Output',
   communityPoolSpend: 'Community Pool Spend',
+  actionLiquidityTournamentVote: 'Liquidity Tournament Vote',
 };
 
 const getLabelForActionCase = (actionCase: ActionView['actionView']['case']): string => {
@@ -144,6 +146,9 @@ export const ActionViewComponent = ({
 
     case 'communityPoolDeposit':
       return <UnimplementedView label='Community Deposit' />;
+
+    case 'actionLiquidityTournamentVote':
+      return <LiquidityTournamentVoteComponent value={actionView.value} />;
 
     default:
       return <UnimplementedView label={getLabelForActionCase(actionView.case)} />;