+++ packages/ui/src/utils/typography.ts
@@ -30,6 +30,8 @@ export const bodyTechnical = cn('font-mono text-textBase font-normal leading-tex
 
 export const small = cn('font-default text-textSm font-normal leading-textXs');
 
+export const smallTechnical = cn('font-mono text-textSm font-normal leading-textSm');
+
 export const detail = cn('font-default text-textXs font-medium leading-textXs');
 
 export const detailTechnical = cn('font-mono text-textXs font-normal leading-textXs');
@@ -61,6 +63,6 @@ export const p = cn('font-default text-textBase font-normal leading-textBase mb-
 
 export const button = cn('font-default text-textBase font-medium leading-textBase');
 
-export const buttonMedium = cn('font-default text-textBase font-medium leading-textBase');
+export const buttonMedium = cn('font-default text-textSm font-medium leading-textBase');
 
-export const buttonSmall = cn('font-default text-textSm font-medium leading-textBase');
+export const buttonSmall = cn('font-default text-textXs font-medium leading-textBase');