+++ packages/bech32m/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @penumbra-zone/bech32
 
+## 14.0.0
+
+### Patch Changes
+
+- Updated dependencies [49ae3ab]
+  - @penumbra-zone/protobuf@8.0.0
+
 ## 13.0.0
 
 ### Patch Changes